examples/mandelbrot/requirements.txt
paho-mqtt>=2.0
numpy>=1.24
matplotlib>=3.7
